Patch from Will Dean (svn@indcomp.co.uk): The attached patch is a bit of a tidy of some of the TMerge code, particularly concentrating on the BEAST of a class CDiffData.
I found the use of !sFilename.IsEmpty() all over the place to mean 'are we
using this file' very confusing (too many double negatives for my feeble
brain), so I did something about that, which lead me into hoisting all the
code which deals with the filenames and descriptive names (sXXXFile and
sXXXName) out into another class called CWorkerFile.

This gets rid of lots of duplicated filename manipulation code in
MainFrm.cpp, and I hope rather cleans things up.

I have taken the two-way and three-way code diff-building code out of the
CDiffData Load function into separate functions.

I would like to move to making these three:

       CFileTextLines                          m_arBaseFile;
       CFileTextLines                          m_arTheirFile;
       CFileTextLines                          m_arYourFile;

local to CDiffData::Load, because their data is not used after the loading
and diffing is complete, and it would save memory not to have redundant
copies of files loaded.    However, they're used outside the class to copy
file-type props across to a saved file, so this still needs a bit of work.

git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@2166 154be670-7969-47af-9ab0-c6fb378d2295