Patch from Dirk (vss2svn at nogga.de):
there is a small bug in the ResText utility that prevents DLGTEMPLATE to be converted correctly.

The definition of the DLGTEMPLATE structure (taken from the MSDN) is:

typedef struct {
   DWORD style;
   DWORD dwExtendedStyle;
   WORD cdit;
   short x;
   short y;
   short cx;
   short cy;
} DLGTEMPLATE, *LPDLGTEMPLATE;

and the style parameter was already parsed in the CountMemReplaceDialogResource. The original code tried to reparse the style parameter after the dwExtendedStyle parameter.


git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@5625 154be670-7969-47af-9ab0-c6fb378d2295