Patch from Andreas Nicolai:
1. week count:
old: The previous implementation took the first date and the last date in the array as time span.

new: The new implementation searches for min and max dates, then aligns the earliest date with a date at the begin of the corresponding week, then this date is stored in a new member variable m_minDate.

2. data gathering:
old: the previous implementation was implemented such that a lot of binary searches (using lower_bound) were executed for _each_ commit. This caused the noticable delay when opening the stats dialog for large number of revisions (e.g. try "Show all" in the TSVN repository and open the stats dialog). Also, reoccuring weeks due to later import of revision histories would be treated as new weeks and thus not giving the correct stats.

new: The new implementation loops over all weeks in the intervals determined in GetWeekCount() and stores for each week/interval the number of commits and file changes per author, it also keeps track of the total commit count and total file change count. At the same time the commits for each author are stored in a mapping. Then a list of author names is created and the list is sorted based on commit count. For that purpose I wrote a binary predicate class MoreCommitsThan to be able to compare authors based on their commit count. As a result, all the sorting during the data gathering is no longer necessary and the time expensive CountCommits() function can be removed alltogether. Further, the required stats are obtained for the min/max author (first and last in the sorted list) and the dialogs statistics can be shown.


git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@10914 154be670-7969-47af-9ab0-c6fb378d2295