To help with debugging the shell extensions: In TortoiseStub, if the current process runs under a debugger and the registry key HKLM\Software\TortoiseSVN\DebugShell is set, load the shell extension dlls from the path of the current process. Then just start the debugger with the TortoiseProc process, open some file-open dialog and the debug shell dlls are loaded, ready for debugging.
Note: the fact that the dlls are now loaded from the path of the process is not a security risk: the registry key must be set in HKLM which no non-admin can do, and the process must be run under a debugger. These conditions can not be set from the outside without the user knowing (if they could, then the system is already compromised).

git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@21273 154be670-7969-47af-9ab0-c6fb378d2295