Patch from Evgeny Kotkov: his patch fixes two issues that may prevent the "fetching tree conflict details" progress dialog from displaying properly.
The first part of the issue is that initializing a CProgressDlg within the
CSVNProgressDlg::ResolvePostOperationConflicts() method requires COM
to be initialized, whereas its calling site currently doesn't do that.

The second part of the issue is that within the same method, the ShowModal()
method is called with the result of GetExplorerHWND() passed as the parent
window handle.  This affects the modality of the progress dialog, and this
part of the issue can be fixed by passing m_hWnd as the parent window
handle.



git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@28207 154be670-7969-47af-9ab0-c6fb378d2295