Patch from Patrick Steinhardt: When you try to commit to a server, but the commit itself is getting rejected because of a hook, then TortoiseSVN will prompt the user to unbreak the lock. Outputting this error message is wrong, though, as rejection through a hook will not cause any locks to be retained after the operation.
The issue stems from the macro `SVN_ERR_IS_UNLOCK_ERROR`, which
TortoiseSVN uses to determine whether unlock instructions shall
be appended to the error messages. Because Subversion allows
pre-lock hooks, an `SVN_ERR_REPOS_HOOK_FAILURE` message is also
considered an unlock error, even though in most cases it is not.
And even if it was an error regarding the pre-lock hook, the user
would not be able to fix that by unlocking something in his
working copy, as the hook is server-side and only used to check
whether a lock shall be granted at all.

Fix the erroneous message by only emitting it if the error is
_not_ an `SVN_ERR_REPOS_HOOK_FAILURE`.

The attached patch does that.

git-svn-id: https://svn.osdn.net/svnroot/tortoisesvn/trunk@28229 154be670-7969-47af-9ab0-c6fb378d2295